requirements.txt
python-twitch-client
forecastiopy
geopy
thesaurus
bs4
pillow
etym
wordcloud
wikipedia
